requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.25.8
python-gitlab==4.4.0
python-dotenv==1.0.1
httpx==0.27.0
gitpython==3.1.43
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
